feat: impl counter using f64

+++ crates/loro-internal/src/op/content.rs
@@ -67,7 +67,7 @@ impl InnerContent {
 #[derive(EnumAsInner, Debug, Clone)]
 pub enum FutureInnerContent {
     #[cfg(feature = "counter")]
-    Counter(i64),
+    Counter(f64),
     Unknown {
         prop: i32,
         value: OwnedValue,
@@ -82,7 +82,7 @@ pub enum RawOpContent<'a> {
     List(ListOp<'a>),
     Tree(TreeOp),
     #[cfg(feature = "counter")]
-    Counter(i64),
+    Counter(f64),
     Unknown {
         prop: i32,
         value: OwnedValue,
